backend/requirements.txt
fastapi[all]==0.111.0
uvicorn[standard]==0.30.1
pydantic-settings==2.3.4
python-dotenv==1.0.1
google-generativeai==0.7.1
psycopg2-binary==2.9.9
pypdf==4.2.0
sentence-transformers==3.0.1

# The conflicting packages, now locked to compatible versions
supabase==2.4.4
langchain==0.2.6
langchain-community==0.2.6
langchain-core==0.2.10
langchain-huggingface==0.0.3